requirements/base.txt
django==4.2.11
djangorestframework==3.15.2
django-filter
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
python-dotenv
pillow

django-cors-headers 


django-celery-beat==2.6.0
redis==5.0.3
celery==5.3.6
flower==2.0.1
